backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
qrcode[pil]==7.4.2
Pillow==10.0.1
Werkzeug==2.3.7
python-decouple==3.8
PyPDF2==3.0.1
Pillow==10.0.1
reportlab==4.0.4